refactor(front): extract logic to a lib
